fixup! feat(dev-infra): tool for staging and publishing releases
Fix issue where next version is incremented when branching off for
a feature-freeze. In extreme rare situations, we might have branched
off for feature-freeze and will branch off for feature-freeze again
once the FF/RC phase ends. In those situations, the version in the
next branch has not been published to NPM at all, and when branching
off for a new feature-freeze/RC, the version _should_ not be bumped
incorrectly as that would mean that we accidentally leave out the
`-next.0` release.