test(dev-infra): move mock git client to shared location
For the release publish tool we created a virtual Git client that
simulates a Git repository at minimal. We want to move it to a
shared location so that we can incrementally add needed functionality
to it, and start using it for newly built tools (otherwise we will
add more and more tools without testing).