feat(dev-infra): add release command for setting NPM dist tag
Introduces a new command for `ng-dev release`, so that the NPM
dist tag can be set for all configured NPM packages.

This command can be useful in case a manual tag needs to
be set. It is primarily used by the release tooling when a
new stable version is cut, and when the previous patch branch
needs to be set as LTS version through a `v{major}-lts` dist tag.

It is necessary to have this as a command so that the release tool
can execute it for old branches where other packages might have been
configured-lts` dist tag.

It is necessary to have this as a command so that the release tool
can execute it for old branches where other packages might have been
configured.